Isolated Plot into its own sole module that uses Vega; removed MUI dependency
