+++ chartlets.js/src/lib/components/Plot/index.ts
@@ -1,0 +1,1 @@
+export { Plot } from "./Plot";
